[TASK] Drop the title field for time slots
Now the begin date always is used as the label in the back end.
