+++ Classes/OldModel/TimeSlot.php
@@ -131,15 +131,4 @@ class Tx_Seminars_OldModel_TimeSlot extends \Tx_Seminars_OldModel_AbstractTimeSp
     {
         return $this->hasRecordPropertyInteger('entry_date');
     }
-
-    /**
-     * Returns an associative array, containing field name/value pairs that need
-     * to be updated in the database. Update means "set the title" so far.
-     *
-     * @return string[] data to update the database entry of the time slot, might be empty
-     */
-    public function getUpdateArray(): array
-    {
-        return ['title' => \html_entity_decode($this->getDate(), ENT_QUOTES | ENT_HTML5, 'utf-8')];
-    }
 }
