+++ Resources/Private/Language/nl.locallang_db.xlf
@@ -740,10 +740,6 @@
                 <source>Time slots</source>
                 <target>Bloktijden</target>
             </trans-unit>
-            <trans-unit id="tx_seminars_timeslots.title" xml:space="preserve" approved="yes">
-                <source>Title</source>
-                <target>Titel</target>
-            </trans-unit>
             <trans-unit id="tx_seminars_timeslots.begin_date" xml:space="preserve" approved="yes">
                 <source>Begin date</source>
                 <target>Begin datum</target>
